requirements.txt
Flask
Authlib
flask-sqlalchemy
flask-migrate
pyodbc
python-dotenv
pytds
sqlalchemy